run unit test with react testing library and jest
