+++ src/logic/___tests___/calculate.test.js
@@ -1,5 +1,4 @@
 import calculate from '../calculate';
-import operate from '../operate';
 
 describe('Test calculate.js file', () => {
   it('Test that 2 + 2 equals 4', () => {
@@ -146,25 +145,3 @@ describe('Test the conditions before the operations in calculate.js', () => {
   });
 });
 
-describe('Tests for operate.js', () => {
-  it('test sum 4 + 5', () => {
-    const result = operate('4', '5', '+');
-    expect(result).toBe('9');
-  });
-  it('test sum 4 - 5', () => {
-    const result = operate('4', '5', '-');
-    expect(result).toBe('-1');
-  });
-  it('test sum 4 * 5', () => {
-    const result = operate('4', '5', 'x');
-    expect(result).toBe('20');
-  });
-  it('test sum 20 / 5', () => {
-    const result = operate('20', '5', '÷');
-    expect(result).toBe('4');
-  });
-  it('test sum 20 % 5', () => {
-    const result = operate('20', '5', '%');
-    expect(result).toBe('0');
-  });
-});
